Tasks and lists
Working hard with task and lists, fixed some complications when changing names on files and folders and moved around. Almost fixed but something is wrong when i create a new task. Im sent to app/tasks/create.php and its all blank but the task updates on the website and in the database.